q_t structure, test bench, misc
The two major changes in this commit are moving from a structure based
solution for the 'q_t' type to a simple arithmetic value type. This
simplifies the interface and is more useful to the user.

A test bench is being added as well, which will replace the manual
checks that had to be done. Instead checks can be read in from a file
and evaluated.

There have been other minor style changes as well.