Removed stdio.h dependency, improved test bench
The dependency on 'stdio.h' has been removed entirely, the library now
uses its own internal numeric conversion routines for input and output
(which are also more generic, even if they are not as well tested).

The test bench has also been improved and extended to cope with more of
the functions provided by the library.

There are now also several command line options that can be passed to
the test program.