Gain, pol2rec, rec2pol, exp, log, deg2rad, rad2deg
A fair amount of functionality has been added in this commit; functions
for calculating the gain in hyperbolic and circular mode (linear mode
gain *should* be one), log and exp, degrees to radians and the reverse,
polar coordinates to rectangular coordinates and the reverse.

More test cases have also been added, and a way of testing the wrapping
mode as opposed to the default behavior of overflow, which is
saturation. There have been some minor style changes (replacing '0' with
'QINT(0)', deleting unused code, hiding an internal function) as well,
and minor miscellaneous changes.