Fixed issue #2, and also some memory issues
A discrepancy within the API has been resolved in favor of consistency
within the other functions as opposed to what is necessarily the best
way. The problem was a function prototype and its implementation had
different type signatures (well, dependent on the size of an "int", on
many platforms they would effectively be the same). A few memory issues
have been resolved as well.